requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
fonttools==4.51.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
scipy==1.13.0
Werkzeug==3.0.3
